requirements.txt
# Render.com için gerekli dosyalar

# requirements.txt (Render için)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==0.28.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
gunicorn==21.2.0
